+++ ProjetEcom/src/model/Vente.java
@@ -12,29 +12,28 @@ import com.google.gson.annotations.Expose;
 @Entity
 public class Vente {
 
-	@Id 
-	@GeneratedValue(strategy=GenerationType.AUTO)
+	@Id
+	@GeneratedValue(strategy = GenerationType.AUTO)
 	@Expose
-	private long id;
+	private long		id;
 	@Expose
-	private int prix;	
-	
-	@ManyToOne 
-	@JoinColumn(name="Livre_id")
-    private Livre livre;
-	
-	@ManyToOne 
-	@JoinColumn(name="Commande_id") 
-    private Commande laCommande;
-    
+	private float		prix;
+
+	@ManyToOne
+	@JoinColumn(name = "Livre_id")
+	private Livre		livre;
+
+	@ManyToOne
+	@JoinColumn(name = "Commande_id")
+	private Commande	laCommande;
+
 	public Vente() {
 	}
 
-	public Vente(int prix) {
+	public Vente(float prix) {
 		super();
 		this.prix = prix;
 	}
-
 
 	public long getId() {
 		return id;
@@ -44,18 +43,18 @@ public class Vente {
 		this.id = id;
 	}
 
-	public int getPrix() {
+	public float getPrix() {
 		return prix;
 	}
 
-	public void setPrix(int prix) {
+	public void setPrix(float prix) {
 		this.prix = prix;
 	}
 
 	public Livre getLivre() {
 		return livre;
 	}
- 
+
 	public void setLivre(Livre livre) {
 		this.livre = livre;
 	}
@@ -64,11 +63,8 @@ public class Vente {
 		return laCommande;
 	}
 
-
 	public void setLaCommande(Commande laCommande) {
 		this.laCommande = laCommande;
 	}
- 
-	
-	
+
 }
